+++ src/cpp/session/include/session/SessionOptions.hpp
@@ -289,6 +289,7 @@ private:
    void resolvePandocPath(const core::FilePath& resourcePath, std::string* pPath);
    void resolveQuartoPath(const core::FilePath& resourcePath, std::string* pPath);
    void resolveCopilotPath(const core::FilePath& resourcePath, std::string* pPath);
+   void resolveNodePath(const core::FilePath& resourcePath, std::string* pPath);
    void resolveRsclangPath(const core::FilePath& resourcePath, std::string* pPath);
 
    void resolveOverlayOptions();